linkscanner-backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
supabase==2.28.2
python-dotenv==1.2.2
pydantic==2.12.1
httpx==0.28.1
requests==2.32.5
PyJWT==2.12.1
python-multipart==0.0.20
